Added CommandCheck function to ReceivePubNub.js
Works the same as in the Android Application.
First we JSONparse the Message into SensorId and ValueId
Which we then sends as parameters into CommandCheck
Then we do If Logic depending on the CommandId value and Value value.

Next step: Logic for showing it inside Home.jsp

Warning: Still have not found out a way to test if we can receive msg.
The only thing i know is that the connection is established.